+++ packages/library-sync/src/commonMain/kotlin/io/realm/kotlin/mongodb/AppConfiguration.kt
@@ -20,6 +20,7 @@ import io.ktor.client.features.logging.Logger
 import io.realm.kotlin.LogConfiguration
 import io.realm.kotlin.Realm
 import io.realm.kotlin.RealmConfiguration
+import io.realm.kotlin.internal.CoreExceptionConverter
 import io.realm.kotlin.internal.RealmLog
 import io.realm.kotlin.internal.interop.sync.MetadataMode
 import io.realm.kotlin.internal.interop.sync.NetworkTransport
@@ -84,6 +85,11 @@ public interface AppConfiguration {
     public class Builder(
         private val appId: String
     ) {
+
+        init {
+            CoreExceptionConverter.initialize()
+        }
+
         private var baseUrl: String = DEFAULT_BASE_URL
         // TODO We should use a multi threaded dispatcher
         //  https://github.com/realm/realm-kotlin/issues/501